Moved new functionality and tests to LetterTransformer
